+++ packages/components/src/index.ts
@@ -1,2 +1,3 @@
 // 集中导出 所有组件
 export * from './tree'
+export * from './utils'