+++ packages/components/src/utils/index.ts
@@ -1,0 +1,3 @@
+// 导出统一定义的接口及类型，便于使用者直接使用
+export * from './random' 
+export * from './tree-utils'